move assets tests to their own test files
